requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
pytest>=6.0
Flask>=2.0.0
opencv-python>=4.5.0
imagehash>=4.3.0
Pillow>=9.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
Flask-Limiter>=3.0.0
cmake>=3.20.0
face_recognition>=1.3.0
reportlab>=3.6.0
SpeechRecognition>=3.8.1
pydub>=0.25.1
python-dotenv>=0.19.0
email-validator>=1.1.0
gunicorn>=20.1.0
google-generativeai>=0.3.0
Flask-Babel>=2.0.0
